refactor(tabs): split request_editor_tabs into per-tab view files
